requirements.txt
Pillow
pyfiglet==0.8.post1
InquirerPy
pytest==6.1.1
prompt-toolkit
tqdm==4.43.0
ttkthemes==3.1.1
termcolor==1.1.0
